Tight Bezier bounds + persistent tessellators + CPU double precision
- Add asymptote/bound.h and bound.cc: extracted from path3.cc
  Provides bound(), boundtri(), cubic curve bound(), Fuzz, maxdepth, run::norm()

- Change TRIPLE from glm::vec3 (float) to glm::dvec3 (double)
  readReal() now returns double; all CPU geometry at full precision

- Add persistent BezierPatch S / BezierTriangle S members to all Bezier classes
  Fast path uses S.Onscreen and S.append() instead of re-tessellating every frame
  Matches Asymptote drawBezierPatch::S -- was root cause of rendering lag

- All Bezier constructors compute tight recursive bounds at construction time
  QueueMesh uses stored Min/Max for offscreen culling (bbox2)

- Fix include ordering: rgba.h/bezierpatch.h before <iostream> to avoid
  symbolmaps.h alternate token conflicts (and/or/not)

+++ v3d-Common/V3dFile/V3dTypes.h
@@ -10,9 +10,9 @@
 using UINT = uint32_t;
 using V3D_BOOL = uint32_t;
 
-using REAL = float;
-using PAIR = glm::vec2;    // REAL  * 2
-using TRIPLE = glm::vec3;  // REAL  * 3
+using REAL = double;
+using PAIR = glm::dvec2;   // REAL  * 2
+using TRIPLE = glm::dvec3; // REAL  * 3
 
 using FLOAT = float;
